Fixing schedulers

+++ app/views/publicListing/companion/BuyNFT.scala
@@ -13,5 +13,5 @@ object BuyNFT {
       constants.FormField.PASSWORD.mapping
     )(Data.apply)(Data.unapply))
 
-  case class Data(publicListingId: String, buyNFTs: Int, gasAmount: Int, gasPrice: Double, password: String)
+  case class Data(publicListingId: String, buyNFTs: Int, gasAmount: Int, gasPrice: BigDecimal, password: String)
 }
